pyproject.toml
[tool.poetry]
name = "hexlet-advanced-testing"
version = "0.1.0"
description = ""
authors = ["akasmall <aka_sm@ro.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.3.2"
pyyaml = "^6.0.2"
faker = "^27.0.0"
schema = "^0.7.7"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
